+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 ## [major.minor.patch] - DD-MM-YYYY
 
+## Breaking Changes
+
+- Deleted `name_filter` from get multiple datasets endpoint [PR#91](https://github.com/quality-match/hari-client/pull/91)
+
 ### New Features
 
-- add method to download media files from dataset [PR#94](https://github.com/quality-match/hari-client/pull/94)
+- Add pagination, sorting and filtering to the get annotation runs, pipelines, AINT learning data, ML annotation models, AI annotation runs methods [PR#91](https://github.com/quality-match/hari-client/pull/91)
+- Add count annotation runs, pipelines, AINT learning data, ML annotation models, AI annotation runs methods [PR#91](https://github.com/quality-match/hari-client/pull/91)
+- Add `ilike` query operator [PR#91](https://github.com/quality-match/hari-client/pull/91)
+
+- Add method to download media files from dataset [PR#94](https://github.com/quality-match/hari-client/pull/94)
 
 #### Consistency in partially failed uploads